+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
  - Optimized WiFi, BLE, mbedTLS, and LWIP buffer sizes for all non-PSRAM configs
  - Split commandline god file into separate organised files
+ - Added a 250ms guard to prevent closing the wardriving view when opening it
 
 ### Fixed
  - Adjusted lockscreen layout for cardputer screens